Added special styling for the header in register.php
As per the comment in register.css:

Since register.php cannot have a user log in and should not have a user register, these sections can be removed.
Because these sections can be removed, the header can be smaller.
Because the header can be smaller, the nav bar should be centered vertically in it.